refactor: move darken terminal to ui module
